fix: update logger parameter in error handling methods and rename test fixtures file
