+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bob-core",
-  "version": "2.0.0-beta.15",
+  "version": "2.0.0-beta.16",
   "description": "BOB Core",
   "type": "module",
   "main": "./dist/cjs/src/index.js",
