+++ src/errors/InvalidOption.ts
@@ -13,11 +13,11 @@ export class InvalidOption extends BobError {
 		super(`Invalid option ${option} in not recognized`);
 	}
 
-	pretty(io: Logger): void {
+	pretty(logger: Logger): void {
 		const options = Object.entries(this.optionsSchema);
 
 		if (options.length > 0) {
-			io.log(`\n${chalk.yellow('Available options')}:`);
+			logger.log(`\n${chalk.yellow('Available options')}:`);
 
 			for (const [name, definition] of options) {
 				const details = getOptionDetails(definition);
@@ -26,11 +26,11 @@ export class InvalidOption extends BobError {
 				const nameWithAlias = `--${name}${alias.length > 0 ? alias.map(a => `, -${a}`).join('') : ''}`;
 				const spaces = ' '.repeat(30 - nameWithAlias.length);
 
-				io.log(`  ${chalk.green(nameWithAlias)} ${spaces} ${details.description || '\b'} ${chalk.white(`(${typeDisplay})`)}`);
+				logger.log(`  ${chalk.green(nameWithAlias)} ${spaces} ${details.description || '\b'} ${chalk.white(`(${typeDisplay})`)}`);
 			}
-			io.log('');
+			logger.log('');
 		}
 
-		io.log(`${chalk.white.bgRed(' ERROR ')} Option ${chalk.bold.yellow(this.option)} is not recognized.`);
+		logger.log(`${chalk.white.bgRed(' ERROR ')} Option ${chalk.bold.yellow(this.option)} is not recognized.`);
 	}
 }
